Mobile RFID-Reader: ID-12 + XBee + Arduino
This is the code for a Mobile RFID-Reader. An ID-12 RFID reader is
controlled by an Arduino. When it detects an RFID card it forwards the
ID of that card to another Arduino via XBee. The project contains two
Arduino sketches, one for the Mobile RFID-Reader and one for the
receiver.